requirements.txt
blinker==1.6.2
click==8.1.7
Faker==19.3.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
